обновление

+++ src/main/java/ru/linapelx/zadachi/zadachaseven/ArrayPI.java
@@ -1,5 +1,5 @@
 package ru.linapelx.zadachi.zadachaseven;
 
 public class ArrayPI {
-    public int[] array = new int[35];
+    public long[] array = new long[35];
 }
