+++ src/main/java/ru/linapelx/zadachi/zadachasix/ArrayPI.java
@@ -1,6 +1,6 @@
 package ru.linapelx.zadachi.zadachasix;
 
 public class ArrayPI {
-    public int[][] arrayA = new int[7][7];
-    public int[][] arrayB = new int[7][7];
+    public long[][] arrayA = new long[7][7];
+    public long[][] arrayB = new long[7][7];
 }
